fix: Disable app locale settin on MIUI due to Xiaomi's broken implementation
Note that this doesn't affect devices running on HyperOS, for more information, see: https://github.com/Pool-Of-Tears/GreenStash/issues/130
