+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2024-05-05T08:54:15.319206505Z">
+        <DropdownSelection timestamp="2024-05-15T13:36:51.260393900Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="PhysicalDevice" identifier="serial=685adbce" />
+              <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\starr\.android\avd\Pixel_4a_API_34.avd" />
             </handle>
           </Target>
         </DropdownSelection>
